pyproject.toml
[tool.poetry]
name = "overthrow"
version = "0.1.0"
description = ""
authors = ["Jan Rydzewski <flegmer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "2.2.*"
psycopg2 = "^2.8.4"
django-environ = "^0.4.5"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
djangorestframework-recaptcha = "^0.2.0"
factory_boy = "^2.12.0"
hypothesis = {extras = ["django"], version = "^5.7.1"}
coverage = "^5.0.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
